+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2.6.1
 
+ - Use delayed block changes in batches to allow large construction without killing the server.
+
 ## 2.6.0
 
  - Show maximum mana in Wand lore.